refactor(reports): extract shared types to reports_types.go
